+++ tests/testthat/test_mcmc.R
@@ -19,24 +19,24 @@ test_that("Mixed", {
 
   testthat::skip_on_cran()
   set.seed(10)
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw"), NA)
   #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "iw"))
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse"), NA)
   #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "diffuse")
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv"), NA)
   #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "csv"))
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv"), NA)
   #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "fsv")
 })
 
@@ -59,24 +59,24 @@ test_that("Quarterly", {
 
   testthat::skip_on_cran()
   set.seed(10)
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw"), NA)
   #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl", variance = "iw"))
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv"), NA)
   #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "csv"))
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv"), NA)
   #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "fsv")
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse"), NA)
   #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "diffuse")
 })
 
@@ -98,24 +98,24 @@ test_that("Monthly", {
 
   testthat::skip_on_cran()
   set.seed(10)
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw"), NA)
   #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl", variance = "iw"))
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv"), NA)
   #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "csv"))
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv"), NA)
   #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "fsv")
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse"), NA)
   #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "diffuse")
 })
 
@@ -139,24 +139,24 @@ test_that("Block exogenous 1", {
 
   testthat::skip_on_cran()
   set.seed(10)
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw"), NA)
   #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "iw"))
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse"), NA)
   #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "diffuse")
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv"), NA)
   #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "csv"))
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv"), NA)
   #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "fsv", a = 1/10)
 })
 
@@ -181,23 +181,66 @@ test_that("Block exogenous 2", {
 
   testthat::skip_on_cran()
   set.seed(10)
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw"), NA)
   #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "iw"))
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse"), NA)
   #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "diffuse")
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv"), NA)
   #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "csv"))
 
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv")
-  mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv")
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv"), NA)
   #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "fsv", a = 1/10)
+})
+
+test_that("Weekly-Monthly MCMC", {
+  set.seed(10237)
+  Y <- matrix(rnorm(400), 100, 4)
+  Y[setdiff(1:100,seq(4, 100, by = 4)), 4] <- NA
+
+  prior_obj <- set_prior(Y = Y, freq = c(rep("w", 3), "m"),
+                         n_lags = 4, n_reps = 10)
+
+  prior_intervals <- matrix(c(
+    -0.5, 0.5,
+    -0.5, 0.5,
+    -0.5, 0.5,
+    -0.5, 0.5), ncol = 2, byrow = TRUE)
+  psi_moments <- interval_to_moments(prior_intervals)
+  prior_psi_mean <- psi_moments$prior_psi_mean
+  prior_psi_Omega <- psi_moments$prior_psi_Omega
+  prior_obj <- update_prior(prior_obj, d = "intercept", prior_psi_mean = prior_psi_mean,
+                            prior_psi_Omega = prior_psi_Omega, n_fcst = 4,
+                            n_fac = 1)
+
+  testthat::skip_on_cran()
+  set.seed(10)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "iw"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "iw"), NA)
+  #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "iw"))
+
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "diffuse"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "diffuse"), NA)
+  #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "diffuse")
+
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "csv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "csv"), NA)
+  #expect_error(mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "csv"))
+
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "minn", variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ss",   variance = "fsv"), NA)
+  expect_error(estimate_mfbvar(mfbvar_prior = prior_obj, prior = "ssng", variance = "fsv"), NA)
+  #mod <- estimate_mfbvar(mfbvar_prior = prior_obj, prior = "dl",   variance = "fsv")
 })